requirements.txt
awscli
boto3
strands-agents
strands-agents-tools 
strands-agents-builder
requests
mcp
streamlit
fastmcp
starlette
python-dotenv
